adding scripts for paper result generation
